Fix HealSign after Bukkit updates for 1.6.1

+++ src/main/java/de/philworld/bukkit/magicsigns/signs/HealSign.java
@@ -39,7 +39,7 @@ public class HealSign extends PurchasableMagicSign {
 	public void onRightClick(PlayerInteractEvent event) {
 		Player p = event.getPlayer();
 
-		int newHealth = p.getHealth() + healAmount;
+		double newHealth = p.getHealth() + healAmount;
 
 		if (newHealth > p.getMaxHealth()) {
 			event.getPlayer().setHealth(p.getMaxHealth());
